float fun

+++ train_RA_CNN.py
@@ -98,7 +98,7 @@ def line_search_train(data_path, wvs_path, documents=None, test_mode=False,
 
     best_so_far = -np.inf
     sent_dropout_star = None
-    for sent_dropout in np.linspace(sent_dropout_range[0], sent_dropout_range[1], num_steps):
+    for sent_dropout in np.linspace(sent_dropout_range[0], sent_dropout_range[1], num_steps, dtype='float32'):
         r_CNN, documents, p, X_doc, y_doc, best_performance = \
             train_CNN_rationales_model(data_path, wvs_path, documents=documents, 
                                 test_mode=test_mode,
